[refactor] 지역 변경 컴포넌트 분리
